+++ tests/test_optimization_data.py
@@ -1,8 +1,8 @@
-from qcio import OptimizationResults
+from qcio import OptimizationData
 
 
 def test_optimization_result_properties(prog_output):
-    opt_res = OptimizationResults(
+    opt_res = OptimizationData(
         trajectory=[prog_output],
     )
 
@@ -21,7 +21,7 @@ def test_optimization_result_properties(prog_output):
 
 
 def test_optimization_save_to_xyz(prog_output, tmp_path):
-    opt_res = OptimizationResults(
+    opt_res = OptimizationData(
         trajectory=[prog_output] * 3,
     )
     opt_res.save(tmp_path / "opt_res.xyz")
@@ -48,9 +48,9 @@ H  0.75016279902412597 -0.33132205318865016 -0.54481406902570462
 
 
 def test_optimization_save_non_xyz(prog_output, tmp_path):
-    opt_res = OptimizationResults(
+    opt_res = OptimizationData(
         trajectory=[prog_output] * 3,
     )
     opt_res.save(tmp_path / "opt_res.json")
-    opt_res_copy = OptimizationResults.open(tmp_path / "opt_res.json")
+    opt_res_copy = OptimizationData.open(tmp_path / "opt_res.json")
     assert opt_res == opt_res_copy