+++ tests/test_serialization.py
@@ -1,4 +1,4 @@
-from qcio import ProgramOutput
+from qcio import Results
 
 
 def test_serialization_to_disk_json(prog_output, tmp_path):
@@ -6,13 +6,13 @@ def test_serialization_to_disk_json(prog_output, tmp_path):
 
     filename = tmp_path / "prog_output.json"
     prog_output.save(filename)
-    reopened = ProgramOutput.open(filename)
+    reopened = Results.open(filename)
     assert prog_output == reopened
 
     # No filename or other extension to json by default
     filename = tmp_path / "prog_output.whatever"
     prog_output.save(filename)
-    reopened = ProgramOutput.open(filename)
+    reopened = Results.open(filename)
     assert prog_output == reopened
 
 
@@ -21,7 +21,7 @@ def test_serialization_to_disk_yaml(prog_output, tmp_path):
     for ext in [".yaml", ".yml"]:
         filename = tmp_path / f"prog_output{ext}"
         prog_output.save(filename)
-        reopened = ProgramOutput.open(filename)
+        reopened = Results.open(filename)
         assert prog_output == reopened
 
 
@@ -30,5 +30,5 @@ def test_serialization_to_disk_toml(prog_output, tmp_path):
 
     filename = tmp_path / "prog_output.toml"
     prog_output.save(filename)
-    reopened = ProgramOutput.open(filename)
+    reopened = Results.open(filename)
     assert prog_output == reopened