+++ tests/test_view.py
@@ -1,4 +1,4 @@
-from qcio import OptimizationResults, ProgramInput, ProgramOutput
+from qcio import CalcInput, OptimizationData, Results
 from qcio.view import generate_optimization_plot
 
 
@@ -7,11 +7,11 @@ def test_generate_optimization_plot_with_single_prog_output_failure(
 ):
     opt_input = prog_input("optimization")
 
-    prog_output = ProgramOutput[ProgramInput, OptimizationResults](
+    prog_output = Results[CalcInput, OptimizationData](
         input_data=opt_input,
         success=False,
         traceback="Traceback...",
-        results=OptimizationResults(trajectory=[prog_output_failure]),
+        data=OptimizationData(trajectory=[prog_output_failure]),
         provenance={"program": "fake-program"},
     )
     generate_optimization_plot(prog_output)